requirements.txt
pandas>=1.5.3
requests>=2.32.3
Flask==3.1.2
Werkzeug==3.1.3
blinker==1.9.0
click==8.2.1
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pip==25.2
python-dotenv==1.1.1